Rename namespace directory to LCDComponents
